Add files via upload
This improves on the previous version. Data are selected using guis. A subset of data are used for bleaching correction. An exponential fit is performed on the peaks from the files selected for bleaching correction. The user is given the option to accept or reject the fit. The data are then analyzed automatically. Outputs a png with various plots and a spreadsheet of all the data.